Unpack actions and action types.
Also, use Jest snapshots to capture expected action structure.
