+++ HelloAgain/actions/friend.js
@@ -1,0 +1,9 @@
+'use strict'
+
+import {
+  UPDATE_FRIEND
+} from './types'
+
+export const updateFriend = (friend) => {
+  return { type: UPDATE_FRIEND, friend: friend }
+}
