+++ HelloAgain/actions/types.js
@@ -1,0 +1,7 @@
+'use strict'
+
+// Contacts
+export const TOGGLE_ACTIVE = 'TOGGLE_ACTIVE'
+
+// Friends
+export const UPDATE_FRIEND = 'UPDATE_FRIEND'
